backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-keycloak==3.9.0
PyJWT==2.8.0
requests==2.31.0